# docs/dev/api/openapi.yaml
openapi: 3.1.0
info:
  title: Auth API
  description: 'Authentication API: accounts, sessions, second factors and client applications. Issues ES256 access tokens and rotating refresh tokens, and publishes the JWKS resource servers verify them with.'
  license:
    name: MIT
  version: 2.0.1
servers:
- url: http://localhost:3000
  description: Local development
paths:
...
  /admin/permissions:
    get:
      tags:
      - admin
      operationId: permissions
      responses:
        '200':
          description: Every permission a role can grant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermissionResponse'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Missing `roles:manage`, or no second factor enrolled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
...
  /auth/personal-access-tokens/exchange:
    post:
      tags:
      - auth
      operationId: exchange
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonalAccessTokenExchangeRequest'
        required: true
      responses:
        '200':
          description: A short-lived access token carrying the token's scopes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalAccessTokenExchangeResponse'
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Unknown, revoked or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Account locked, suspended or inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '413':
          description: Body larger than 64 KB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '415':
          description: Body is not JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /users/me/tokens:
    get:
      tags:
      - account
      operationId: list
      responses:
        '200':
          description: Personal access tokens that still work, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonalAccessTokenResponse'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: '`first_party_session_required`: a token delegated to a client or issued for a personal access token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
    post:
      tags:
      - account
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePersonalAccessTokenRequest'
        required: true
      responses:
        '201':
          description: Token created; its secret is in this response only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedPersonalAccessTokenResponse'
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Recent re-authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: '`too_many_tokens`'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '413':
          description: Body larger than 64 KB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '415':
          description: Body is not JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Invalid name or lifetime, or a scope the account does not hold
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
  /users/me/tokens/{id}:
    delete:
      tags:
      - account
      operationId: revoke
      parameters:
      - name: id
        in: path
        description: Token id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Token revoked
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: '`first_party_session_required`: a token delegated to a client or issued for a personal access token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: No such token for this account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
...
components:
  schemas:
...
    CreatePersonalAccessTokenRequest:
      type: object
      required:
      - name
      properties:
        expires_in_days:
          type:
          - integer
          - 'null'
          format: int64
          description: 1 to 365 (default 90).
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
          description: |-
            Permissions the token's access tokens carry; each must be held by the
            account. Empty: none.
...
    CreatedPersonalAccessTokenResponse:
      allOf:
      - $ref: '#/components/schemas/PersonalAccessTokenResponse'
      - type: object
        required:
        - secret
        properties:
          secret:
            type: string
            description: The secret, shown once.
...
    ErrorBody:
      type: object
      description: Body of every error response.
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: 'Stable and machine-readable: `invalid_credentials`, `reauthentication_required`, ...'
        message:
          type: string
          description: Human-readable explanation; may change between versions.
...
    PermissionResponse:
      type: object
      required:
      - name
      properties:
        description:
          type:
          - string
          - 'null'
        name:
          type: string
          description: '`resource:action`.'
    PersonalAccessTokenExchangeRequest:
      type: object
      required:
      - token
      properties:
        token:
          type: string
          description: A personal access token (`aapat_...`).
    PersonalAccessTokenExchangeResponse:
      type: object
      required:
      - access_token
      - token_type
      - expires_in
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
          format: int64
          description: Seconds.
          minimum: 0
        token_type:
          type: string
          description: Always `Bearer`.
    PersonalAccessTokenResponse:
      type: object
      required:
      - id
      - name
      - scopes
      - created_at
      - expires_at
      properties:
        created_at:
          type: integer
          format: int64
        expires_at:
          type: integer
          format: int64
        id:
          type: string
          format: uuid
        last_used_at:
          type:
          - integer
          - 'null'
          format: int64
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...